admin/requirements.txt
streamlit
supabase
python-dotenv
plotly
pandas
openpyxl
requests
